[seller] refactor: 중복 정의된 getAccountVerification을 entity/account로 통합
