+++ apps/seller/src/entity/account/account.type.ts
@@ -1,0 +1,9 @@
+export interface AccountVerificationDetail {
+  id: number
+  sellerId: number
+  bankCode: string
+  accountNumber: string
+  accountHolder: string
+  verified: boolean
+  createdAt: string
+}
