+++ apps/seller/src/entity/account/index.ts
@@ -1,0 +1,2 @@
+export { getAccountVerification } from './account.api'
+export type { AccountVerificationDetail } from './account.type'
