+++ apps/seller/src/entity/register/register.api.ts
@@ -1,10 +1,7 @@
-import { isAxiosError } from 'axios'
-
 import type { ApiResponse } from '@/entity/auth/types'
 import { client } from '@/shared/utils/axios'
 
 import type {
-  AccountVerificationDetail,
   AccountVerificationRequest,
   AccountVerificationResult,
   RegisterDocumentsRequest,
@@ -41,18 +38,6 @@ export async function getStoreNames(
   }
 
   return data.result
-}
-
-export async function getAccountVerification(): Promise<AccountVerificationDetail | null> {
-  try {
-    const { data } = await client.get<ApiResponse<AccountVerificationDetail>>(
-      '/api/v1/seller/sellers/account-verifications',
-    )
-    return data.result ?? null
-  } catch (err) {
-    if (isAxiosError(err) && err.response?.status === 400) return null
-    throw err
-  }
 }
 
 export async function verifyAccount(
